index: inform user of successful settings injection
The esptool-js code logs the following confusing message:

Not changing the image

We have changed the message to be less confusing:

Not changing image flash parameters (size, mode, frequency)

But to make it even more clear that the Willow settings have actually
been injected into the image, let's add explicit logging for that. Move
the calls to generate NVS partition and inject it into the dist image
into the try catch block, and write a message to the terminal after
injecting.