index.js: hide reset button after stop
The stop action can be slow. As it can appear to be doing nothing,
people might be tempted to click reset after clicking stop. When timed
right, this throws a bunch of errors in the browser console.

Hide the reset button after clicking stop to avoid this, and restore it
after the disconnect action has been completed.